+++ tests/test_strspn_s.c
@@ -18,7 +18,7 @@ int test_strspn_s (void)
 {
     errno_t rc = 0;
     rsize_t count = 0;
-    int32_t std_count = 0;
+    int std_count = 0;
 
 
 /*--------------------------------------------------*/
@@ -31,7 +31,7 @@ int test_strspn_s (void)
 
     if (count != 0) {
         printf("%s %u  count=%d  std_count=%d  rc=%d \n",
-                     __FUNCTION__, __LINE__,  count, std_count, rc);
+                     __FUNCTION__, __LINE__,  (int)count, std_count, rc);
     }
 
 /*--------------------------------------------------*/
@@ -44,7 +44,7 @@ int test_strspn_s (void)
 
     if (count != 0) {
         printf("%s %u  count=%d  std_count=%d  rc=%d \n",
-                     __FUNCTION__, __LINE__,  count, std_count, rc);
+                     __FUNCTION__, __LINE__,  (int)count, std_count, rc);
     }
 
 /*--------------------------------------------------*/
@@ -65,7 +65,7 @@ int test_strspn_s (void)
 
     if (count != 0) {
         printf("%s %u  count=%d  std_count=%d  rc=%d \n",
-                     __FUNCTION__, __LINE__,  count, std_count, rc);
+                     __FUNCTION__, __LINE__,  (int)count, std_count, rc);
     }
 
 /*--------------------------------------------------*/
@@ -78,7 +78,7 @@ int test_strspn_s (void)
 
     if (count != 0) {
         printf("%s %u  count=%d  std_count=%d  rc=%d \n",
-                     __FUNCTION__, __LINE__,  count, std_count, rc);
+                     __FUNCTION__, __LINE__,  (int)count, std_count, rc);
     }
 
 /*--------------------------------------------------*/
@@ -91,7 +91,7 @@ int test_strspn_s (void)
 
     if (count != 0) {
         printf("%s %u  count=%d  std_count=%d  rc=%d \n",
-                     __FUNCTION__, __LINE__,  count, std_count, rc);
+                     __FUNCTION__, __LINE__,  (int)count, std_count, rc);
     }
 
 /*--------------------------------------------------*/
@@ -104,7 +104,7 @@ int test_strspn_s (void)
 
     if (count != 0) {
         printf("%s %u  count=%d  std_count=%d  rc=%d \n",
-                     __FUNCTION__, __LINE__,  count, std_count, rc);
+                     __FUNCTION__, __LINE__,  (int)count, std_count, rc);
     }
 
 /*--------------------------------------------------*/
@@ -120,7 +120,7 @@ int test_strspn_s (void)
 
     if (count != 0) {
         printf("%s %u  Error  count=%d rc=%d \n",
-                     __FUNCTION__, __LINE__, count, rc);
+               __FUNCTION__, __LINE__, (int)count, rc);
     }
 
 /*--------------------------------------------------*/
@@ -136,7 +136,7 @@ int test_strspn_s (void)
 
     if (count != 1) {
         printf("%s %u  count=%d  std_count=%d  rc=%d \n",
-                     __FUNCTION__, __LINE__,  count, std_count, rc);
+                     __FUNCTION__, __LINE__,  (int)count, std_count, rc);
     }
 
 /*--------------------------------------------------*/
@@ -152,7 +152,7 @@ int test_strspn_s (void)
 
     if (count != 2) {
         printf("%s %u  count=%d  std_count=%d  rc=%d \n",
-                     __FUNCTION__, __LINE__,  count, std_count, rc);
+                     __FUNCTION__, __LINE__,  (int)count, std_count, rc);
     }
 
 /*--------------------------------------------------*/
@@ -168,7 +168,7 @@ int test_strspn_s (void)
 
     if (count != 3) {
         printf("%s %u  count=%d  std_count=%d  rc=%d \n",
-                     __FUNCTION__, __LINE__,  count, std_count, rc);
+                     __FUNCTION__, __LINE__,  (int)count, std_count, rc);
     }
 
 /*--------------------------------------------------*/
@@ -185,7 +185,7 @@ int test_strspn_s (void)
     std_count = strspn(str1, str2);
     if (count != std_count) {
         printf("%s %u  count=%d  std_count=%d  rc=%d \n",
-                     __FUNCTION__, __LINE__,  count, std_count, rc);
+                     __FUNCTION__, __LINE__,  (int)count, std_count, rc);
     }
 
 /*--------------------------------------------------*/
@@ -202,7 +202,7 @@ int test_strspn_s (void)
     std_count = strspn(str1, str2);
     if (count != std_count) {
         printf("%s %u  count=%d  std_count=%d  rc=%d \n",
-                     __FUNCTION__, __LINE__,  count, std_count, rc);
+                     __FUNCTION__, __LINE__,  (int)count, std_count, rc);
     }
 
 
@@ -220,7 +220,7 @@ int test_strspn_s (void)
     std_count = strspn(str1, str2);
     if (count != std_count) {
         printf("%s %u  count=%d  std_count=%d  rc=%d \n",
-                     __FUNCTION__, __LINE__,  count, std_count, rc);
+                     __FUNCTION__, __LINE__,  (int)count, std_count, rc);
     }
 
 /*--------------------------------------------------*/
@@ -237,7 +237,7 @@ int test_strspn_s (void)
     std_count = strspn(str1, str2);
     if (count != std_count) {
         printf("%s %u  count=%d  std_count=%d  rc=%d \n",
-                     __FUNCTION__, __LINE__,  count, std_count, rc);
+                     __FUNCTION__, __LINE__,  (int)count, std_count, rc);
     }
 
 /*--------------------------------------------------*/
@@ -254,7 +254,7 @@ int test_strspn_s (void)
     std_count = strspn(str1, str2);
     if (count != std_count) {
         printf("%s %u  count=%d  std_count=%d  rc=%d \n",
-                     __FUNCTION__, __LINE__,  count, std_count, rc);
+                     __FUNCTION__, __LINE__,  (int)count, std_count, rc);
     }
 
 /*--------------------------------------------------*/
@@ -271,7 +271,7 @@ int test_strspn_s (void)
     std_count = strspn(str1, str2);
     if (count != std_count) {
         printf("%s %u  count=%d  std_count=%d  rc=%d \n",
-                     __FUNCTION__, __LINE__,  count, std_count, rc);
+                     __FUNCTION__, __LINE__,  (int)count, std_count, rc);
     }
 
 /*--------------------------------------------------*/
